semantic: impl `From<LiteralType>` for `Type`
